drawing more boundaries
