+++ server/src/auth/auth.data-access.ts
@@ -1,0 +1,7 @@
+import { Students } from '../../db/schema';
+
+export const insertNewStudent = async (studentDoc) => {
+  const newStudent = new Students(studentDoc);
+
+  await newStudent.save();
+};
