+++ server/src/profile/profile.data-access.ts
@@ -1,19 +1,8 @@
 import { Reviews, Students } from '../../db/schema';
 
-export const findStudent = async (netId: string) => {
-  const student = await Students.findOne({ netId }).exec();
-  return student;
-};
-
 export const findReviewDoc = async (reviewId: string) => {
   const review = await Reviews.findOne({ _id: reviewId }).exec();
   return review;
-};
-
-export const insertNewStudent = async (studentDoc) => {
-  const newStudent = new Students(studentDoc);
-
-  await newStudent.save();
 };
 
 export const updateStudentReviews = async (
