+++ server/src/review/review.data-access.ts
@@ -36,18 +36,3 @@ export const updateReviewLikes = async (
     { $pull: { likedBy: netId } },
   );
 };
-
-export const updateReviewVisibility = async (
-  reviewId: string,
-  reported: number,
-  visible: number,
-) => {
-  await Reviews.updateOne(
-    { _id: reviewId },
-    { $set: { visible: visible, reported: reported } },
-  );
-};
-
-export const removeReview = async (reviewId: string) => {
-  await Reviews.remove({ _id: reviewId });
-};
