+++ server/tests/Search.test.ts
@@ -11,8 +11,8 @@ beforeAll(async () => {
       firstName: 'John',
       lastName: 'Smith',
       netId: 'js0',
-      affiliation: null,
-      token: null,
+      affiliation: '',
+      token: '',
       privilege: 'regular',
       reviews: [],
       likedReviews: [],
